To assist with cartridge/assembler output analysis, added two additional/supporting command-line arguments:
* `--cartoffsets`: used with `-g`, will calculate byte-based offsets based on v32
cartridge base
* `--vbinoffsets`: used with `-g`, will calculate byte-based offsets based on VBIN
datafile base

The current default output of a `-g` debug file is Vircon32 word-based in Vircon32
cartridge memory.

Depending on the nature of the investigation (ie doing raw hexdumps of the
cartridge or vbin file on the computer running Vircon32), the offsets will not
immediately be useful.

Essentially just looked up the existing `-g` functionality and replicated based on
it.