Addressing a call to deprecated function `sprintf` by instead using the recommended `snprintf`.
Size set to 23, which appears to be the maximum (indeed only) length of the generated string.

`DesktopEmulator/Emulator/GUI.cpp:679:13: warning: 's
printf' is deprecated: This function is provided for compatibility reasons only.  Due to security conc
erns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.
[-Wdeprecated-declarations]`