Conditional checking and selection of GTK library, checking 4, then 3, then defaulting to 2 if none are found.
Some data had to be moved around to ensure it existed in CMakeLists.txt (like EMULATOR_SRC)

With the updated logic in CMakeLists.txt, the appropriate ExternalLibraries/osdialog/osdialog_gtkX.c file is
able to be utilized.

Preprocessor checks for OSDIALOG_GTK4 have also been added.

The GTK4 API is rather different, so it does not currently compile (based off the cleanly working GTK3 code under GTK3).

It seems the notion of dialogs and widgets has changed enough in GTK4 where there will need to be bigger changes to the code.